流式布局添加新内容拉伸的功能，实现了#15的功能。

+++ MyLayout/Lib/MyFlowLayout.m
@@ -373,7 +373,7 @@ bestSingleLineArray:(NSMutableArray*)bestSingleLineArray
     }
     
     
-    for (int i = index; i < sbs.count; i++) {
+    for (NSInteger i = index; i < sbs.count; i++) {
         
         
         NSMutableArray *calcArray2 = [NSMutableArray arrayWithArray:calcArray];
